Made Version 16
Please download version 16 of the Rubik's Cube Timer (zipped) or just the file!

+++ Uninstall Rubiks Cube Timer Application.py
@@ -1,29 +1,26 @@
 import shutil
-import os
 import time
 print("")
-
-current_file_path = str(os.path.realpath(__file__))
-
-try:
+i = input("Are you sure you want to uninstall 'Rubik's Cube Timer'? Enter 'Yes' to continue or press to enter key to exit: ")
+if i == "Yes":
     try:
-        shutil.rmtree("C:/Rubik's Cube Timer")
-        print("The application has been uninstalled, please delete the Rubik's Cube Timer v12.exe file.")
+        try:
+            shutil.rmtree("C:/Rubik's Cube Timer")
+        except FileNotFoundError:
+            print("")
+            print("Already removed!")
+        time.sleep(2)
         print("")
-        time.sleep(2.5)
-        input("Press the 'enter' key to finish uninstalling the application. ")
-        os.remove(current_file_path)
-    except FileNotFoundError:
-        print("The application is already uninstalled, please delete the Rubik's Cube Timer.exe file.")
+        print("------------------------------")
+        input("Almost done! Press the enter key to finish uninstalling: ")
+    except:
         print("")
-        time.sleep(2.5)
-        input("Press the 'enter' key to finish uninstalling the application. ")
-except:
+        print("------------------------------")
+        print("Error")
+        print("")
+        print("Unable to remove application files.")
+        print("Please contact developer.")
+        time.sleep(2)
+        input("Press the enter key to exit uninstaller: ")
+elif i == "":
     print("")
-    print("------------------------------")
-    print("Error")
-    print("")
-    print("Unable to remove files for application.")
-    print("Please contact developer.")
-    time.sleep(2.5)
-    input("Press the enter key to exit setup: ")
